+++ src/ListBroking/AppBundle/EventListener/AjaxAuthenticationExceptionListener.php
@@ -1,0 +1,28 @@
+<?php
+/**
+ * @author     Samuel Castro <samuel.castro@adclick.pt>
+ * @copyright  2015 Adclick
+ */
+
+namespace ListBroking\AppBundle\EventListener;
+
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Security\Core\Exception\AuthenticationException;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
+use Symfony\Component\HttpKernel\Event\GetResponseForExceptionEvent;
+
+class AjaxAuthenticationExceptionListener
+{
+    public function onCoreException(GetResponseForExceptionEvent $event)
+    {
+        $exception = $event->getException();
+        $request = $event->getRequest();
+
+        if ($request->isXmlHttpRequest()) {
+            if ($exception instanceof AuthenticationException || $exception instanceof AccessDeniedException) {
+                $event->setResponse(new Response('', 403));
+            }
+        }
+    }
+
+}